Moved code to PreferenceActivity
