+++ app/src/main/java/com/amaze/filemanager/ui/activities/AboutActivity.java
@@ -24,8 +24,7 @@ import static com.amaze.filemanager.utils.Utils.openURL;
 
 import com.amaze.filemanager.LogHelper;
 import com.amaze.filemanager.R;
-import com.amaze.filemanager.ui.activities.superclasses.BasicActivity;
-import com.amaze.filemanager.ui.fragments.preference_fragments.PreferencesConstants;
+import com.amaze.filemanager.ui.activities.superclasses.PreferenceActivity;
 import com.amaze.filemanager.ui.theme.AppTheme;
 import com.amaze.filemanager.utils.Billing;
 import com.amaze.filemanager.utils.Utils;
@@ -35,7 +34,6 @@ import com.google.android.material.snackbar.Snackbar;
 import com.mikepenz.aboutlibraries.Libs;
 import com.mikepenz.aboutlibraries.LibsBuilder;
 
-import android.app.Activity;
 import android.content.SharedPreferences;
 import android.content.res.Configuration;
 import android.graphics.Bitmap;
@@ -47,7 +45,6 @@ import android.view.MenuItem;
 import android.view.View;
 import android.widget.TextView;
 
-import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.appcompat.widget.Toolbar;
 import androidx.coordinatorlayout.widget.CoordinatorLayout;
@@ -55,7 +52,7 @@ import androidx.palette.graphics.Palette;
 import androidx.preference.PreferenceManager;
 
 /** Created by vishal on 27/7/16. */
-public class AboutActivity extends BasicActivity implements View.OnClickListener {
+public class AboutActivity extends PreferenceActivity implements View.OnClickListener {
 
   private static final String TAG = "AboutActivity";
 
@@ -88,7 +85,7 @@ public class AboutActivity extends BasicActivity implements View.OnClickListener
   private static final String URL_REPO_RATE = "market://details?id=com.amaze.filemanager";
 
   @Override
-  protected void onCreate(@Nullable Bundle savedInstanceState) {
+  public void onCreate(@Nullable Bundle savedInstanceState) {
     super.onCreate(savedInstanceState);
 
     if (getAppTheme().equals(AppTheme.DARK)) {
@@ -146,38 +143,6 @@ public class AboutActivity extends BasicActivity implements View.OnClickListener
         (v, hasFocus) -> {
           mAppBarLayout.setExpanded(hasFocus, true);
         });
-  }
-
-  @Override
-  public void onConfigurationChanged(@NonNull Configuration newConfig) {
-    super.onConfigurationChanged(newConfig);
-
-    int currentNightMode = newConfig.uiMode & Configuration.UI_MODE_NIGHT_MASK;
-
-    if (AppTheme.getTheme(
-            Integer.parseInt(mSharedPref.getString(PreferencesConstants.FRAGMENT_THEME, "4")))
-        .equals(AppTheme.SYSTEM))
-      switch (currentNightMode) {
-        case Configuration.UI_MODE_NIGHT_NO:
-          getUtilsProvider().getThemeManager().setAppTheme(AppTheme.getTheme(0));
-          restartPC(this);
-          break;
-        case Configuration.UI_MODE_NIGHT_YES:
-          getUtilsProvider().getThemeManager().setAppTheme(AppTheme.getTheme(1));
-          restartPC(this);
-          break;
-      }
-  }
-
-  public static void restartPC(final Activity activity) {
-    if (activity == null) return;
-
-    final int enter_anim = android.R.anim.fade_in;
-    final int exit_anim = android.R.anim.fade_out;
-    activity.overridePendingTransition(enter_anim, exit_anim);
-    activity.finish();
-    activity.overridePendingTransition(enter_anim, exit_anim);
-    activity.startActivity(activity.getIntent());
   }
 
   /**
