+++ app/src/main/java/com/amaze/filemanager/ui/activities/MainActivity.java
@@ -115,7 +115,6 @@ import com.amaze.filemanager.ui.fragments.SearchWorkerFragment;
 import com.amaze.filemanager.ui.fragments.TabFragment;
 import com.amaze.filemanager.ui.fragments.preference_fragments.PreferencesConstants;
 import com.amaze.filemanager.ui.strings.StorageNamingHelper;
-import com.amaze.filemanager.ui.theme.AppTheme;
 import com.amaze.filemanager.ui.views.appbar.AppBar;
 import com.amaze.filemanager.ui.views.drawer.Drawer;
 import com.amaze.filemanager.utils.AppConstants;
@@ -1225,33 +1224,6 @@ public class MainActivity extends PermissionsActivity
     super.onConfigurationChanged(newConfig);
     // Pass any configuration change to the drawer toggls
     drawer.onConfigurationChanged(newConfig);
-
-    int currentNightMode = newConfig.uiMode & Configuration.UI_MODE_NIGHT_MASK;
-
-    if (AppTheme.getTheme(
-            Integer.parseInt(getPrefs().getString(PreferencesConstants.FRAGMENT_THEME, "4")))
-        .equals(AppTheme.SYSTEM))
-      switch (currentNightMode) {
-        case Configuration.UI_MODE_NIGHT_NO:
-          getUtilsProvider().getThemeManager().setAppTheme(AppTheme.getTheme(0));
-          restartPC(this);
-          break;
-        case Configuration.UI_MODE_NIGHT_YES:
-          getUtilsProvider().getThemeManager().setAppTheme(AppTheme.getTheme(1));
-          restartPC(this);
-          break;
-      }
-  }
-
-  public static void restartPC(final Activity activity) {
-    if (activity == null) return;
-
-    final int enter_anim = android.R.anim.fade_in;
-    final int exit_anim = android.R.anim.fade_out;
-    activity.overridePendingTransition(enter_anim, exit_anim);
-    activity.finish();
-    activity.overridePendingTransition(enter_anim, exit_anim);
-    activity.startActivity(activity.getIntent());
   }
 
   @Override
